13 Jan 2021

+++ ATLists/ATLists.Sql/SqlCategory.cs
@@ -11,9 +11,9 @@ namespace ATLists.SQL
         public string CategoryType { get; set; }
 
 
-        public byte SqlEntryStorage { get; set; }
-        public byte SqlColorable { get; set; }
-        public byte SqlIconable { get; set; }
-        public byte SqlSingleText { get; set; }
+        public int SqlEntryStorage { get; set; }
+        public int SqlColorable { get; set; }
+        public int SqlIconable { get; set; }
+        public int SqlSingleText { get; set; }
     }
 }
