+++ ATLists/ATLists.Sql/SqlEntry.cs
@@ -11,11 +11,11 @@ namespace ATLists.SQL
         public string EntryType { get; set; }
 
 
-        public byte SqlColorable { get; set; }
-        public byte SqlIconable { get; set; }
-        public byte SqlMultyText { get; set; }
-        public byte SqlQualifiable { get; set; }
-        public byte SqlQuantifiable { get; set; }
-        public byte SqlSingleText { get; set; }
+        public int SqlColorable { get; set; }
+        public int SqlIconable { get; set; }
+        public int SqlMultyText { get; set; }
+        public int SqlQualifiable { get; set; }
+        public int SqlQuantifiable { get; set; }
+        public int SqlSingleText { get; set; }
     }
 }
